Mount multiple ceph.conf and keyring as secrets
This change introduces a new way of ceph.{conf,keyring} processing,
reducing the amount of knowledge that the operator should have when
a ceph cluster is configured as backend. In particular a new
'CephSecret' key has been added and it represents a list of secrets
passed to the operator.
The operator will mount those files to the /etc/ceph target directory
and the inline config should be passed to point to one of the existing
clusters.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>