Add Glance Version detect Job
This patch introduces a glance-detect-version job. The goal of this job
is to analyze the glance container image and determine whether Glance
should be run in ProxyPass or mod_wsgi.
We need this new deployment step to make sure that human operators
coming from a previous releases do not see Glance failing when they
upgrade the -operator, which is now able to deploy the Image service
only using mod_wsgi.
This issue originated due to the incompatibility with older Glance images
where mod_wsgi is not present (the rpm is missing).
When a human operators perform a minor update, they usually update the
glance-operator first, then the edit the openstackVersion CR to accept
the new available release. During this time, if the new glance-operator
is run with an older Glance image, it results in the service
CrashLoopBackOff.
However, instead of forcing a rollout of a new GlanceAPI version during
the operators upgrade stage (it would simply represent a workaround for
this problem), the idea is to enhance the glance-operator and create a
compatibility layer to either use a deployment method (proxypass vs
mod_wsgi) or another based on the source image.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>